+++ src/addin/client-interfaces/addin-box-control.ts
@@ -1,0 +1,14 @@
+/**
+ * Interface for defining a control for Box add-ins
+ */
+export interface AddinBoxControl {
+  /**
+   * The display name of the control
+   */
+  name: string;
+
+  /**
+   * The action of the control
+   */
+  action: string;
+}